requirements.txt
fastapi[standard]
uvicorn
motor
bcrypt
python-jose[cryptography]
passlib
pydantic[email]
pydantic
pydantic-settings
pyjwt
openai
pinecone-client
scikit-learn
pdfplumber
python-docx
tiktoken
langchain_community
langchain
langchain-openai
OpenAI
python-dotenv
websocket
spacy
pinecone-client[grpc]
